plat/rpi4: fix incorrect serial baudrate with recent start4.elf
Ref: https://developer.trustedfirmware.org/T605

Since the introduction of the 2019.11.18 start4.elf by the Pi Foundation, it
appears that the hardcoded clock frequency (PLAT_RPI4_VPU_CLK_RATE) that
serves as the base clock against which the serial clock rate divisors are set,
has doubled in value. This means that, if you are using TF-A with a recent
start4.elf, the serial output will be garbled when read at 155200 bauds.

Testing shows that, whereas the base clock rate that needs to be fed to
rpi3_console_init() hasn't changed, the divisor values we read have doubled
compared to their 2019.11.18 start4.elf counterparts, which indicates that the
internal clock value they are indexed on must have doubled too.

This patch fixes this issue by updating PLAT_RPI4_VPU_CLK_RATE to its most
recent value. Unfortunately, due to the proprietary nature of the SoC, we
don't have a way of reading this value dynamically (or a way to read the
version of start4.elf to set it accordingly), which prevents us from providing
a patch that allows baudrate preservation with older versions of start4.elf.
In order to get serial output from TF-A with a pre 2019.11.18 start4.elf, you
therefore need to set your baudrate to 57600 instead of 115200.

We also take this opportunity to fix the code to use the full 24-bit divisor,
rather the 12 bit integer integer part, as this allows the baudrate to be
properly set when using a 'core_freq' that is not 250 or 500 in 'config.txt'.

Signed-off-by: Pete Batard <pete@akeo.ie>